% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraCE21ka_functions.R
\name{downscaleTrace}
\alias{downscaleTrace}
\title{Title}
\usage{
downscaleTrace(
  i,
  new.data.list,
  var.names,
  y1.list,
  y2.list,
  lonLim,
  latLim,
  hist.trace,
  data,
  model,
  local.var,
  trace.model.var.names,
  global.nc.attributes
)
}
\arguments{
\item{i}{TBW}

\item{new.data.list}{TBW}

\item{var.names}{TBW}

\item{y1.list}{TBW}

\item{y2.list}{TBW}

\item{lonLim}{TBW}

\item{latLim}{TBW}

\item{hist.trace}{TBW}

\item{data}{TBW}

\item{model}{TBW}

\item{local.var}{TBW}

\item{trace.model.var.names}{TBW}

\item{global.nc.attributes}{TBW}
}
\value{
TBW
}
\description{
Title
}
\examples{
#TBW
}
