% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraCE21ka_functions.R
\name{loadManualTrace}
\alias{loadManualTrace}
\title{Title}
\usage{
loadManualTrace(
  file_list,
  var_list,
  trace.y1,
  trace.y2,
  lonLim,
  latLim,
  dictionary = system.file("extdata", "TraCE21ka_dictionary.csv", package =
    "TraCE21kaDSR"),
  var_selection = trace.model.var.names,
  compute_wss = TRUE
)
}
\arguments{
\item{file_list}{TBW}

\item{var_list}{TBW}

\item{trace.y1}{TBW}

\item{trace.y2}{TBW}

\item{lonLim}{TBW}

\item{latLim}{TBW}

\item{dictionary}{TBW}

\item{var_selection}{TBW}

\item{compute_wss}{TBW}
}
\value{
TBW
}
\description{
Title
}
\examples{
#TBW
}
