% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraCE21ka_functions.R
\name{loadManualTraceData}
\alias{loadManualTraceData}
\title{Title}
\usage{
loadManualTraceData(
  file,
  var,
  trace.y1,
  trace.y2,
  lonLim = trace.lon,
  latLim = trace.lat,
  dictionary = system.file("extdata", "TraCE21ka_dictionary.csv", package =
    "TraCE21kaDSR")
)
}
\arguments{
\item{file}{TBW}

\item{var}{TBW}

\item{trace.y1}{TBW}

\item{trace.y2}{TBW}

\item{lonLim}{TBW}

\item{latLim}{TBW}

\item{dictionary}{TBW}
}
\value{
TBW
}
\description{
Title
}
\examples{
#TBW
}
