% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraCE21ka_functions.R
\name{loadTrace}
\alias{loadTrace}
\title{Title}
\usage{
loadTrace(
  file_list,
  var_list,
  lonLim = trace.lon,
  latLim = trace.lat,
  start_date = "1551-01-01",
  end_date = "1990-12-31",
  years = 1961:1990,
  dictionary = system.file("extdata", "TraCE21ka_dictionary.csv", package =
    "TraCE21kaDSR"),
  var_selection = trace.model.var.names,
  compute_wss = TRUE
)
}
\arguments{
\item{file_list}{TBW}

\item{var_list}{TBW}

\item{lonLim}{TBW}

\item{latLim}{TBW}

\item{start_date}{TBW}

\item{end_date}{TBW}

\item{years}{TBW}

\item{dictionary}{TBW}

\item{var_selection}{TBW}

\item{compute_wss}{TBW}
}
\value{
TBW
}
\description{
Title
}
\examples{
#TBW
}
