% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraCE21ka_functions.R
\name{loadTraceData}
\alias{loadTraceData}
\title{Title}
\usage{
loadTraceData(
  file,
  var = NULL,
  lonLim = trace.lon,
  latLim = trace.lat,
  start_date = "1551-01-01",
  end_date = "1990-12-31",
  years = 1961:1990,
  dictionary = system.file("extdata", "TraCE21ka_dictionary.csv", package =
    "TraCE21kaDSR")
)
}
\arguments{
\item{file}{TBW}

\item{var}{TBW}

\item{lonLim}{TBW}

\item{latLim}{TBW}

\item{start_date}{TBW}

\item{end_date}{TBW}

\item{years}{TBW}

\item{dictionary}{TBW}
}
\value{
TBW
}
\description{
Title
}
\examples{
#TBW
}
