% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UERRA_functions.R
\name{loadUerra}
\alias{loadUerra}
\title{Load UERRA dataset in climate4R grid format}
\usage{
loadUerra(
  file,
  var,
  lonLim = uerra.lon,
  latLim = uerra.lat,
  dictionary = system.file("extdata", "UERRA_dictionary.csv", package = "TraCE21kaDSR"),
  new.coordinates
)
}
\arguments{
\item{file}{TBW}

\item{var}{TBW}

\item{lonLim}{TBW}

\item{latLim}{TBW}

\item{dictionary}{TBW}

\item{new.coordinates}{TBW}
}
\value{
TBW
}
\description{
Load UERRA dataset in climate4R grid format
}
\examples{
# TBW
}
