% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraCE21ka_functions.R
\name{traceFileNames}
\alias{traceFileNames}
\title{Title}
\usage{
traceFileNames(indir, vars = NULL, file.number = 1)
}
\arguments{
\item{indir}{TBW}

\item{vars}{TBW}

\item{file.number}{TBW}
}
\value{
TBW
}
\description{
Title
}
\examples{
 #TBW
}
