% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraCE21ka_functions.R
\name{downscaleTrace}
\alias{downscaleTrace}
\title{Title}
\usage{
downscaleTrace(
  files_n,
  outdir,
  trace_dir,
  vars = NULL,
  lonLim = c(-25, 25),
  latLim = c(25, 50),
  hist.trace,
  mod_data,
  model,
  selection_vars = NULL,
  global.nc.attributes = NULL
)
}
\arguments{
\item{files_n}{TBW}

\item{outdir}{TBW}

\item{trace_dir}{TBW}

\item{vars}{TBW}

\item{lonLim}{TBW}

\item{latLim}{TBW}

\item{hist.trace}{TBW}

\item{mod_data}{TBW}

\item{model}{TBW}

\item{selection_vars}{TBW}

\item{global.nc.attributes}{TBW}
}
\value{
TBW
}
\description{
Title
}
\examples{
#TBW
}
